migrate with_test_home to common
